requirements.txt

# ========================
# 🌐 FastAPI y servidor ASGI
# ========================
fastapi==0.115.6
uvicorn==0.34.0
starlette==0.41.3

# ========================
# 🧾 Validación y Tipado
# ========================
pydantic==2.10.4
pydantic_core==2.27.2
typing_extensions==4.12.2
annotated-types==0.7.0

# ========================
# 🗂️ Middlewares y utilidades web
# ========================
python-multipart==0.0.6      # Subida de archivos / formularios
python-dotenv==1.1.1         # Manejo de variables de entorno

# ========================
# 🛡️ Seguridad / Autenticación
# ========================
passlib==1.7.4               # Hashing de contraseñas
bcrypt==4.3.0                # Algoritmo de hashing
python-jose==3.5.0           # JWT
rsa==4.9.1                   # Criptografía RSA
cryptography==45.0.5         # Dependencia de python-jose
ecdsa==0.19.1                # Firmas digitales
pyasn1==0.6.1                # ASN.1 (tokens)

# ========================
# 🐘 Base de datos (PostgreSQL + ORM)
# ========================
SQLAlchemy==2.0.41
psycopg2-binary==2.9.10
greenlet==3.2.3              # Requerido por SQLAlchemy

# ========================
# ⚙️ Internas y dependencias auxiliares
# ========================
anyio==4.7.0
h11==0.14.0
sniffio==1.3.1
exceptiongroup==1.2.2
cffi==1.17.1
pycparser==2.22
six==1.17.0
idna==3.10

# ========================
# 🧰 Herramientas CLI y utilidades
# ========================
click==8.1.7
colorama==0.4.6

# ========================
# 🧪 Testing
# ========================
pytest
httpx

